backend/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Database
mysql-connector-python==8.3.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
python-dateutil==2.8.2
requests==2.31.0

# Async Operations
aiofiles==23.2.1
httpx==0.26.0

# Redis (for caching and session management)
redis==5.0.1

# Validation & Serialization
email-validator==2.1.0

# LLM Integration
openai==1.54.3
anthropic==0.39.0

# Document Parsing
PyPDF2==3.0.1
python-docx==1.1.2
markdown==3.7
beautifulsoup4==4.12.3
lxml==5.3.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx==0.26.0  # For TestClient
